+++ lib/gui/app.js
@@ -35,13 +35,13 @@ const s3Packages = require('../shared/s3-packages');
 const release = require('../shared/release');
 const store = require('../shared/store');
 const packageJSON = require('../../package.json');
-const flashState = require('./models/flash-state');
+const flashState = require('../shared/models/flash-state');
 const settings = require('./models/settings');
 const windowProgress = require('./os/window-progress');
 const analytics = require('./modules/analytics');
 const updateNotifier = require('./components/update-notifier');
-const availableDrives = require('./models/available-drives');
-const selectionState = require('./models/selection-state');
+const availableDrives = require('../shared/models/available-drives');
+const selectionState = require('../shared/models/selection-state');
 const driveScanner = require('./modules/drive-scanner');
 const osDialog = require('./os/dialog');
 const exceptionReporter = require('./modules/exception-reporter');
