+++ lib/gui/components/drive-selector/controllers/drive-selector.js
@@ -21,8 +21,8 @@ const _ = require('lodash');
 const messages = require('../../../../shared/messages');
 const constraints = require('../../../../shared/drive-constraints');
 const analytics = require('../../../modules/analytics');
-const availableDrives = require('../../../models/available-drives');
-const selectionState = require('../../../models/selection-state');
+const availableDrives = require('../../../../shared/models/available-drives');
+const selectionState = require('../../../../shared/models/selection-state');
 
 module.exports = function(
   $q,
