+++ lib/gui/modules/image-writer.js
@@ -24,7 +24,7 @@ const angular = require('angular');
 const _ = require('lodash');
 const childWriter = require('../../child-writer');
 const settings = require('../models/settings');
-const flashState = require('../models/flash-state');
+const flashState = require('../../shared/models/flash-state');
 const windowProgress = require('../os/window-progress');
 const analytics = require('../modules/analytics');
