+++ lib/gui/pages/finish/controllers/finish.js
@@ -17,8 +17,8 @@
 'use strict';
 
 const settings = require('../../../models/settings');
-const flashState = require('../../../models/flash-state');
-const selectionState = require('../../../models/selection-state');
+const flashState = require('../../../../shared/models/flash-state');
+const selectionState = require('../../../../shared/models/selection-state');
 const analytics = require('../../../modules/analytics');
 
 module.exports = function($state) {
