+++ lib/gui/pages/main/controllers/drive-selection.js
@@ -17,7 +17,7 @@
 'use strict';
 
 const settings = require('../../../models/settings');
-const selectionState = require('../../../models/selection-state');
+const selectionState = require('../../../../shared/models/selection-state');
 const analytics = require('../../../modules/analytics');
 const exceptionReporter = require('../../../modules/exception-reporter');
