+++ lib/gui/pages/main/controllers/flash.js
@@ -18,7 +18,7 @@
 
 const messages = require('../../../../shared/messages');
 const settings = require('../../../models/settings');
-const flashState = require('../../../models/flash-state');
+const flashState = require('../../../../shared/models/flash-state');
 const driveScanner = require('../../../modules/drive-scanner');
 const utils = require('../../../../shared/utils');
 const notification = require('../../../os/notification');
