+++ lib/gui/pages/main/controllers/main.js
@@ -17,11 +17,11 @@
 'use strict';
 
 const settings = require('../../../models/settings');
-const flashState = require('../../../models/flash-state');
+const flashState = require('../../../../shared/models/flash-state');
 const analytics = require('../../../modules/analytics');
 const exceptionReporter = require('../../../modules/exception-reporter');
-const availableDrives = require('../../../models/available-drives');
-const selectionState = require('../../../models/selection-state');
+const availableDrives = require('../../../../shared/models/available-drives');
+const selectionState = require('../../../../shared/models/selection-state');
 
 module.exports = function(
   TooltipModalService,
