+++ lib/shared/models/flash-state.js
@@ -16,13 +16,9 @@
 
 'use strict';
 
-/**
- * @module Etcher.Models.FlashState
- */
-
 const _ = require('lodash');
-const store = require('../../shared/store');
-const units = require('../../shared/units');
+const store = require('../store');
+const units = require('../units');
 
 /**
  * @summary Reset flash state
