+++ lib/shared/models/selection-state.js
@@ -17,7 +17,7 @@
 'use strict';
 
 const _ = require('lodash');
-const store = require('../../shared/store');
+const store = require('../store');
 const availableDrives = require('./available-drives');
 
 /**
