+++ tests/gui/pages/main.spec.js
@@ -6,9 +6,9 @@ const path = require('path');
 const supportedFormats = require('../../../lib/shared/supported-formats');
 const angular = require('angular');
 const settings = require('../../../lib/gui/models/settings');
-const flashState = require('../../../lib/gui/models/flash-state');
-const availableDrives = require('../../../lib/gui/models/available-drives');
-const selectionState = require('../../../lib/gui/models/selection-state');
+const flashState = require('../../../lib/shared/models/flash-state');
+const availableDrives = require('../../../lib/shared/models/available-drives');
+const selectionState = require('../../../lib/shared/models/selection-state');
 require('angular-mocks');
 
 describe('Browser: MainPage', function() {
