+++ tests/shared/models/selection-state.spec.js
@@ -1,12 +1,28 @@
+/*
+ * Copyright 2016 resin.io
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 'use strict';
 
 const m = require('mochainon');
 const _ = require('lodash');
 const path = require('path');
-const availableDrives = require('../../../lib/gui/models/available-drives');
-const selectionState = require('../../../lib/gui/models/selection-state');
+const availableDrives = require('../../../lib/shared/models/available-drives');
+const selectionState = require('../../../lib/shared/models/selection-state');
 
-describe('Browser: selectionState', function() {
+describe('Model: selectionState', function() {
 
   describe('given a clean state', function() {
